Ãrepos,services, Local creation Form

+++ src/main/java/com/ufr/tlib/dataManagementServices/IUserService.java
@@ -10,7 +10,7 @@ public interface IUserService {
     public void updateUser(User user) throws DataIntegrityViolationException;
     public void updateProfil(User user) throws UserNotFoundException;
     public void deleteUserById(int id) throws UserNotFoundException;
-    public User getUserById(int id) throws UserNotFoundException;
+    public User getUserById(long id) throws UserNotFoundException;
 }
